+++ src/main/java/br/ufpa/labes/spm/repository/impl/agent/EmailConfigurationDAO.java
@@ -4,7 +4,7 @@ import br.ufpa.labes.spm.repository.impl.BaseDAO;
 import br.ufpa.labes.spm.repository.interfaces.agent.IEmailConfigurationDAO;
 import br.ufpa.labes.spm.domain.EmailConfiguration;
 
-public class EmailConfigurationDAO extends BaseDAO<EmailConfiguration, Integer>
+public class EmailConfigurationDAO extends BaseDAO<EmailConfiguration, Long>
     implements IEmailConfigurationDAO {
 
   protected EmailConfigurationDAO(Class<EmailConfiguration> businessClass) {
