+++ src/main/java/br/ufpa/labes/spm/repository/impl/connections/BranchConCondToMultipleConDAO.java
@@ -4,7 +4,7 @@ import br.ufpa.labes.spm.repository.impl.BaseDAO;
 import br.ufpa.labes.spm.repository.interfaces.connections.IBranchConCondToMultipleConDAO;
 import br.ufpa.labes.spm.domain.BranchConCondToMultipleCon;
 
-public class BranchConCondToMultipleConDAO extends BaseDAO<BranchConCondToMultipleCon, Integer>
+public class BranchConCondToMultipleConDAO extends BaseDAO<BranchConCondToMultipleCon, Long>
     implements IBranchConCondToMultipleConDAO {
 
   protected BranchConCondToMultipleConDAO(Class<BranchConCondToMultipleCon> businessClass) {
