+++ src/main/java/br/ufpa/labes/spm/repository/impl/connections/DependencyDAO.java
@@ -4,7 +4,7 @@ import br.ufpa.labes.spm.repository.impl.BaseDAO;
 import br.ufpa.labes.spm.repository.interfaces.connections.IDependencyDAO;
 import br.ufpa.labes.spm.domain.Dependency;
 
-public class DependencyDAO extends BaseDAO<Dependency, Integer> implements IDependencyDAO {
+public class DependencyDAO extends BaseDAO<Dependency, Long> implements IDependencyDAO {
 
   protected DependencyDAO(Class<Dependency> businessClass) {
     super(businessClass);
