+++ src/main/java/br/ufpa/labes/spm/repository/impl/log/EventDAO.java
@@ -4,7 +4,7 @@ import br.ufpa.labes.spm.repository.impl.BaseDAO;
 import br.ufpa.labes.spm.repository.interfaces.log.IEventDAO;
 import br.ufpa.labes.spm.domain.Event;
 
-public class EventDAO extends BaseDAO<Event, Integer> implements IEventDAO {
+public class EventDAO extends BaseDAO<Event, Long> implements IEventDAO {
 
   protected EventDAO(Class<Event> businessClass) {
     super(businessClass);
