+++ src/main/java/br/ufpa/labes/spm/repository/impl/log/ResourceEventDAO.java
@@ -4,7 +4,7 @@ import br.ufpa.labes.spm.repository.impl.BaseDAO;
 import br.ufpa.labes.spm.repository.interfaces.log.IResourceEventDAO;
 import br.ufpa.labes.spm.domain.ResourceEvent;
 
-public class ResourceEventDAO extends BaseDAO<ResourceEvent, Integer> implements IResourceEventDAO {
+public class ResourceEventDAO extends BaseDAO<ResourceEvent, Long> implements IResourceEventDAO {
 
   protected ResourceEventDAO(Class<ResourceEvent> businessClass) {
     super(businessClass);
