+++ src/main/java/br/ufpa/labes/spm/repository/impl/plainActivities/EnactionDescriptionDAO.java
@@ -4,7 +4,7 @@ import br.ufpa.labes.spm.repository.impl.BaseDAO;
 import br.ufpa.labes.spm.repository.interfaces.plainActivities.IEnactionDescriptionDAO;
 import br.ufpa.labes.spm.domain.EnactionDescription;
 
-public class EnactionDescriptionDAO extends BaseDAO<EnactionDescription, Integer>
+public class EnactionDescriptionDAO extends BaseDAO<EnactionDescription, Long>
     implements IEnactionDescriptionDAO {
 
   protected EnactionDescriptionDAO(Class<EnactionDescription> businessClass) {
