+++ src/main/java/br/ufpa/labes/spm/repository/impl/plainActivities/ParameterDAO.java
@@ -4,7 +4,7 @@ import br.ufpa.labes.spm.repository.impl.BaseDAO;
 import br.ufpa.labes.spm.repository.interfaces.plainActivities.IParameterDAO;
 import br.ufpa.labes.spm.domain.Parameter;
 
-public class ParameterDAO extends BaseDAO<Parameter, Integer> implements IParameterDAO {
+public class ParameterDAO extends BaseDAO<Parameter, Long> implements IParameterDAO {
 
   protected ParameterDAO(Class<Parameter> businessClass) {
     super(businessClass);
