+++ src/main/java/br/ufpa/labes/spm/repository/impl/plannerInfo/AgentInstSuggestionToAgentDAO.java
@@ -4,7 +4,7 @@ import br.ufpa.labes.spm.repository.impl.BaseDAO;
 import br.ufpa.labes.spm.repository.interfaces.plannerInfo.IAgentInstSuggestionToAgentDAO;
 import br.ufpa.labes.spm.domain.AgentInstSuggestionToAgent;
 
-public class AgentInstSuggestionToAgentDAO extends BaseDAO<AgentInstSuggestionToAgent, Integer>
+public class AgentInstSuggestionToAgentDAO extends BaseDAO<AgentInstSuggestionToAgent, Long>
     implements IAgentInstSuggestionToAgentDAO {
 
   protected AgentInstSuggestionToAgentDAO(Class<AgentInstSuggestionToAgent> businessClass) {
